refactor: move ecstore rebalance metadata helpers
